+++ pkg/reporting/types.go
@@ -79,53 +79,3 @@ func (d Distribution) String() string {
 	}
 	return "unknown"
 }
-
-type InstanceReport struct {
-	Events []InstanceReportEvent `json:"events"`
-}
-
-type InstanceReportEvent struct {
-	ReportedAt                int64  `json:"reported_at"`
-	LicenseID                 string `json:"license_id"`
-	InstanceID                string `json:"instance_id"`
-	ClusterID                 string `json:"cluster_id"`
-	AppStatus                 string `json:"app_status"`
-	IsKurl                    bool   `json:"is_kurl"`
-	KurlNodeCountTotal        int    `json:"kurl_node_count_total"`
-	KurlNodeCountReady        int    `json:"kurl_node_count_ready"`
-	K8sVersion                string `json:"k8s_version"`
-	K8sDistribution           string `json:"k8s_distribution,omitempty"`
-	KotsVersion               string `json:"kots_version"`
-	KotsInstallID             string `json:"kots_install_id,omitempty"`
-	KurlInstallID             string `json:"kurl_install_id,omitempty"`
-	IsGitOpsEnabled           bool   `json:"is_gitops_enabled"`
-	GitOpsProvider            string `json:"gitops_provider"`
-	DownstreamChannelID       string `json:"downstream_channel_id,omitempty"`
-	DownstreamChannelSequence uint64 `json:"downstream_channel_sequence,omitempty"`
-	DownstreamChannelName     string `json:"downstream_channel_name,omitempty"`
-	DownstreamSequence        *int64 `json:"downstream_sequence,omitempty"`
-	DownstreamSource          string `json:"downstream_source,omitempty"`
-	InstallStatus             string `json:"install_status,omitempty"`
-	PreflightState            string `json:"preflight_state,omitempty"`
-	SkipPreflights            bool   `json:"skip_preflights"`
-	ReplHelmInstalls          int    `json:"repl_helm_installs"`
-	NativeHelmInstalls        int    `json:"native_helm_installs"`
-}
-
-type PreflightReport struct {
-	Events []PreflightReportEvent `json:"events"`
-}
-
-type PreflightReportEvent struct {
-	ReportedAt      int64  `json:"reported_at"`
-	LicenseID       string `json:"license_id"`
-	InstanceID      string `json:"instance_id"`
-	ClusterID       string `json:"cluster_id"`
-	Sequence        int64  `json:"sequence"`
-	SkipPreflights  bool   `json:"skip_preflights"`
-	InstallStatus   string `json:"install_status"`
-	IsCLI           bool   `json:"is_cli"`
-	PreflightStatus string `json:"preflight_status"`
-	AppStatus       string `json:"app_status"`
-	KotsVersion     string `json:"kots_version"`
-}
